requirements.txt

Flask==2.3.0
flask-cors==4.0.0
Flask-JWT-Extended==4.4.4
bcrypt==4.0.1
python-dotenv==1.0.0
pymysql==1.1.0
